requirementst.txt
asgiref==3.5.2
autoflake==1.4
autopep8==1.6.0
black==22.6.0
click==8.1.3
Django==4.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djhtml==1.5.2
flake8==5.0.3
isort==5.10.1
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.0
pyflakes==2.5.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
